add css selector for .header-anchor for mdx header link styling
Automatically parsed mdx header links were being styled with a selector for all `anchor` tags, which was including explicitely placed links with []() syntax. This changes to select only for `.header-anchor` when applying styling for left-sided chain-link icon, keeping it only visible on hover, but allows other links to render normally.