Update index.md
The way the code in this tutorial is written, 1 token is equal to 1 Wei, not 1 Ether. If I send one Ether (1000000000000000000 Wei) to the buy() method, I would recieve 1000000000000000000 tokens, not 1 token.

The wording of the sentence confused me and made me think there was a bug somewhere in the code or in my local development environment.